File renaming with first author last name, year except for mix of l5pc channels and hh channels
