+++ jaxley_mech/channels/__init__.py
@@ -1,1 +1,12 @@
-from . import aoyama00, benison01, chen24, fm97, hh, kwu09, l5pc, lk04, tfmc90, uiki96
+from . import (
+    aoyama00,
+    benison01,
+    chen24,
+    fohlmeister97,
+    hh,
+    kamiyama09,
+    l5pc,
+    liu04,
+    torre90,
+    usui96,
+)
